+++ search-app/src/cmr/search/data/temporal_ranges_to_elastic.clj
@@ -1,13 +1,16 @@
 (ns cmr.search.data.temporal-ranges-to-elastic
-  "Functions to "
+  "Functions to convert temporal range search params into an elastic sort script to sort by
+  temporal overlap for temporal relevancy"
  (:require
   [clj-time.coerce :as time-coerce]
   [clj-time.core :as time]
+  [clojure.java.io :as io]
   [cmr.common.util :as util]
   [cmr.search.services.query-walkers.temporal-range-extractor :as temporal-range-extractor]))
 
 (def default-temporal-start-date-time
-  "The default date-time to use for relevancy calculations when one is not specified in the temporal range"
+  "The default date-time to use for relevancy calculations when one is not specified in the
+  temporal range. This is the same default used by umm-spec."
   (time-coerce/to-long (time/date-time 1970 1 1)))
 
 (defn temporal-range->elastic-param
@@ -26,20 +29,12 @@
     (util/map-keys->snake_case temporal-range)))
 
 (def script
-  "def totalOverlap = 0;
-   for (range in temporalRanges)
-   {
-     def overlapStartDate = range.start_date;
-     if (doc['start-date'].value != 0 && doc['start-date'].value > overlapStartDate)
-      { overlapStartDate = doc['start-date'].value; }
-     def overlapEndDate = range.end_date;
-     if (doc['end-date'].value != 0 && doc['end-date'].value < overlapEndDate)
-      { overlapEndDate = doc['end-date'].value; }
-     if (overlapEndDate > overlapStartDate)
-      { totalOverlap += overlapEndDate - overlapStartDate; }
-   }
-   if (rangeSpan > 0) { totalOverlap / rangeSpan; }
-   else { 0; }")
+  "Groovy script used by elastic to calculate temporal overlap based on the collection's indexed
+  start and end date and the temporal ranges specified as search params. The temporal overlap is
+  the percentage of the ranges covered by the collection.
+  For each temporal range search param, calculate span of time that the collection overlaps. Add
+  all of the spans and divide by the sum of the time spans in the temporal range search params."
+  (slurp (io/resource "temporal_overlap.groovy")))
 
 (defn temporal-overlap-sort-script
  "Create the script to sort by temporal overlap percent in descending order. Get the temporal ranges