+++ system-int-test/test/cmr/system_int_test/search/collection_relevancy/collection_relevancy.clj
@@ -28,6 +28,7 @@
 
 (deftest relevancy-temporal-ranges
   (dev-sys-util/eval-in-dev-sys `(query-to-elastic/set-sort-use-relevancy-score! true))
+  (dev-sys-util/eval-in-dev-sys `(query-to-elastic/set-sort-use-temporal-relevancy! true))
   (let [coll1 (d/ingest "PROV1" (dc/collection {:short-name "Usage-30"
                                                 :version-id "2"
                                                 :entry-title "Elevation coll1"